# samples/spring-boot-3/customer-service/src/main/java/io/github/blueprintplatform/samples/customerservice/api/dto/CustomerCreateRequest.java
package io.github.blueprintplatform.samples.customerservice.api.dto;

import jakarta.validation.constraints.Email;
import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.Size;

public record CustomerCreateRequest(
    @NotBlank @Size(min = 2, max = 80) String name, @NotBlank @Email String email) {}