# samples/spring-boot-4/customer-service-client/src/main/resources/customer-api-docs.yaml
openapi: 3.1.0
info:
  title: Customer Service API
  description: Customer Service API with type-safe generic responses using OpenAPI
  version: 0.8.3
servers:
  - url: http://localhost:8085/customer-service
    description: Local service URL
paths:
  /customers/{customerId}:
    get:
      tags:
        - customer-controller
      operationId: getCustomer
      parameters:
        - name: customerId
          in: path
          required: true
          schema:
            type: integer
            format: int32
            minimum: 1
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceResponseCustomerDto"
    put:
      tags:
        - customer-controller
      operationId: updateCustomer
      parameters:
        - name: customerId
          in: path
          required: true
          schema:
            type: integer
            format: int32
            minimum: 1
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerUpdateRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceResponseCustomerDto"
    delete:
      tags:
        - customer-controller
      operationId: deleteCustomer
      parameters:
        - name: customerId
          in: path
          required: true
          schema:
            type: integer
            format: int32
            minimum: 1
      responses:
        "200":
          description: OK
  /customers:
    get:
      tags:
        - customer-controller
      operationId: getCustomers
      parameters:
        - name: name
          in: query
          required: false
          schema:
            type: string
        - name: email
          in: query
          required: false
          schema:
            type: string
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 0
            minimum: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 5
            maximum: 10
            minimum: 1
        - name: sortBy
          in: query
          required: false
          schema:
            type: string
            default: customerId
            enum:
              - customerId
              - name
              - email
        - name: direction
          in: query
          required: false
          schema:
            type: string
            default: ASC
            enum:
              - ASC
              - DESC
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceResponsePageCustomerDto"
    post:
      tags:
        - customer-controller
      operationId: createCustomer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerCreateRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceResponseCustomerDto"
components:
  schemas:
    CustomerUpdateRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 80
          minLength: 2
        email:
          type: string
          format: email
          minLength: 1
      required:
        - email
        - name
    CustomerDto:
      type: object
      properties:
        customerId:
          type: integer
          format: int32
        name:
          type: string
        email:
          type: string
    Meta:
      type: object
      properties:
        serverTime:
          type: string
          format: date-time
        sort:
          type: array
          items:
            $ref: "#/components/schemas/Sort"
      x-ignore-model: true
    ServiceResponseCustomerDto:
      allOf:
        - $ref: "#/components/schemas/ServiceResponse"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/CustomerDto"
      x-api-wrapper: true
      x-api-wrapper-datatype: CustomerDto
    Sort:
      type: object
      properties:
        field:
          type: string
        direction:
          type: string
          enum:
            - ASC
            - DESC
      x-ignore-model: true
    CustomerCreateRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 80
          minLength: 2
        email:
          type: string
          format: email
          minLength: 1
      required:
        - email
        - name
    PageCustomerDto:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/CustomerDto"
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        hasNext:
          type: boolean
        hasPrev:
          type: boolean
      x-ignore-model: true
    ServiceResponsePageCustomerDto:
      allOf:
        - $ref: "#/components/schemas/ServiceResponse"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/PageCustomerDto"
      x-api-wrapper: true
      x-api-wrapper-datatype: PageCustomerDto
      x-data-container: Page
      x-data-item: CustomerDto
    ServiceResponse:
      type: object
      properties:
        data:
          type: object
        meta:
          $ref: "#/components/schemas/Meta"
      required:
        - meta
      x-ignore-model: true
    ServiceResponseVoid:
      type: object
      properties:
        data:
          type: object
        meta:
          $ref: "#/components/schemas/Meta"
      required:
        - meta
      x-ignore-model: true